refactor: 패키지 분리 및 BoardFactory 상수 분리
