+++ src/main/java/chess/domain/board/Board.java
@@ -1,4 +1,7 @@
-package chess.domain;
+package chess.domain.board;
+
+import chess.domain.piece.Piece;
+import chess.domain.Position;
 
 import java.util.Map;
