+++ src/main/java/chess/domain/piece/ColorType.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.piece;
 
 public enum ColorType {
